JOBSBATCH-123 refactoring: extract file upload related code
