+++ smartling-job-batches-api/src/main/java/com/smartling/api/jobbatches/util/FileUploadClientFactory.java
@@ -1,4 +1,4 @@
-package com.smartling.api.jobbatches.v2;
+package com.smartling.api.jobbatches.util;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.smartling.api.v2.client.ClientFactory;
